add mem verification

+++ camera/metadata/src/metadata_utils.cpp
@@ -508,7 +508,7 @@ std::shared_ptr<CameraMetadata> MetadataUtils::DecodeFromString(std::string sett
     IF_COND_PRINT_MSG_AND_RETURN(ret != EOK,
         "MetadataUtils::DecodeFromString Failed to copy memory for metadata header")
     IF_COND_PRINT_MSG_AND_RETURN(meta->items_start >= actualMemSize
-        || (actualMemSize - meta->items_start) < meta->item_count * itemLen,
+        || (actualMemSize - meta->items_start) < (uint64_t)meta->item_count * itemLen,
         "MetadataUtils::DecodeFromString invalid item_start")
     decodeData += headerLength;
     camera_metadata_item_entry_t *item = GetMetadataItems(meta);
